AI ROaster

+++ TJ Notebook/frontend/src/context/notes/NoteState.jsx
@@ -1,32 +1,41 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import NoteContext from "./NoteContext";
 import Cookies from 'universal-cookie';
 
-const NoteState = (props) => {
-  const host="http://localhost:8000";
+export default function NoteState(props) {
+  const host="https://tj-notebook.vercel.app";
   const [notes, setNotes] = useState([]);
+  const [progress,setProgress]=useState(0);
+  
+  const cookies = new Cookies();
   //Fetching All Notes
   const getAllNotes=async ()=>{
     try{
+        const authtoken=cookies.get('auth-token');
+        if(authtoken){
+          setProgress(90);
+        // console.log("GetAllNotes",authtoken);
       const response=await fetch(`${host}/api/notes/fetchall`,{
         method:'GET',
         headers:{
           'Content-Type':'application/json',
-          'auth-token':'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoiNjYwMmM1NmE4MzY5MjFhNGY1ZGMwZDRjIn0sImlhdCI6MTcxMTcyNzcwOH0.V35Yba3VrX6Ixo6NNDYepj8rViqkuNda0K3wp9DYzro'
+          'auth-token':authtoken
         }
       });
       setNotes(await response.json());
+      setProgress(100);
+    }
     }
     catch(error){
-      console.log(error);
+      // console.log(error);
     }
   }
-  const cookies = new Cookies();
-  const authtoken=cookies.get('auth-token');
 
   //Add a Note
   const addNote = async (note) => {
     try{
+      setProgress(90);
+    const authtoken=cookies.get('auth-token');
     const response=await fetch(`${host}/api/notes/addnote`,{
       method:'POST',
       headers:{
@@ -36,17 +45,16 @@ const NoteState = (props) => {
       body:JSON.stringify(note)
     });
     const data=await response.json();
-    console.log("Adding a new Note");
+    // console.log("Adding a new Note");
     const created_note = {
       "_id": `${data._id}`,
       "user": `${data.user}`,
       "title": `${note.title}`,
       "description": `${note.description}`,
-      "tag": `${note.tag}`,
-      "date": "2024-03-27T11:37:16.550Z",
-      "__v": 0
+      "tag": `${note.tag}`
     };
     setNotes(notes.concat(created_note));
+    setProgress(100);
   }
   catch(error){
     console.error(error);
@@ -55,7 +63,13 @@ const NoteState = (props) => {
 
   //Deleting A Note using Its ID
   const deleteNote = async (id) => {
-    console.log(id);
+const authtoken=cookies.get('auth-token');
+if(window.confirm("Are your Sure you want to delete this Note")){
+  setProgress(90);
+  const newNotes = notes.filter((note) => {
+    return note._id != id;
+  });
+  setNotes(newNotes);
     const response=await fetch(`${host}/api/notes/deletenote/${id}`,{
       method:'DELETE',
       headers:{
@@ -63,17 +77,17 @@ const NoteState = (props) => {
         'auth-token':authtoken
       }
     });
+    setProgress(100);
     // console.log("Deleting Note with id", id);
-    const newNotes = notes.filter((note) => {
-      return note._id != id;
-    });
-    setNotes(newNotes);
+  }
   }
 
   //Editing A Note using its ID
   const editNote = async (id, title, description, tag) => {
     //TODO API Call
+    const authtoken=cookies.get('auth-token');
     try{
+    setProgress(90);
     const response=await fetch(`${host}/api/notes/updatenote/${id}`,{
       method:'PUT',
       headers:{
@@ -95,7 +109,7 @@ const NoteState = (props) => {
       }
     }
     setNotes(newNotes);
-
+    setProgress(100);
   }
   catch(error){
     console.error(error);
@@ -103,9 +117,8 @@ const NoteState = (props) => {
   }
 
   return (
-    <NoteContext.Provider value={{ notes, setNotes, addNote, deleteNote, editNote ,getAllNotes}}>
+    <NoteContext.Provider value={{ notes, setNotes, addNote, deleteNote, editNote ,getAllNotes,progress,setProgress}}>
       {props.children}
     </NoteContext.Provider>
   )
 }
-export default NoteState;
